pyproject.toml
[project]
name = "charging-system-b2b"
version = "0.1.0"
description = ""
authors = [
    {name = "mh bakhshipour",email = "mh.bakhshipour@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "markdown (>=3.10,<4.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "environs (>=14.5.0,<15.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redlock (>=1.2.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
